moved queue-invites to their own files since the GET endpoint needs to be public
