+++ packages/server/src/queue/queue.controller.ts
@@ -256,59 +256,4 @@ export class QueueController {
       return;
     }
   }
-
-  /**
-   * Creates a new queue invite for the given queue
-   */
-  @Post(':queueId/invite')
-  @Roles(Role.TA, Role.PROFESSOR)
-  async createQueueInvite(
-    @Param('queueId', ParseIntPipe) queueId: number,
-    @Res() res: Response,
-  ): Promise<Response<void>> {
-    try {
-      await this.queueService.createQueueInvite(queueId);
-      res.status(HttpStatus.CREATED).send();
-      return;
-    } catch (err) {
-      throw err;
-    }
-  }
-
-  /**
-   * Deletes a queue invite for the given queue
-   */
-  @Delete(':queueId/invite')
-  @Roles(Role.TA, Role.PROFESSOR)
-  async deleteQueueInvite(
-    @Param('queueId', ParseIntPipe) queueId: number,
-    @Res() res: Response,
-  ): Promise<Response<void>> {
-    try {
-      await this.queueService.deleteQueueInvite(queueId);
-      res.status(HttpStatus.NO_CONTENT).send(); // NO_CONTENT is the correct status code for a successful DELETE request
-      return;
-    } catch (err) {
-      throw err;
-    }
-  }
-
-  /**
-   * Edits a queue invite for the given queue
-   */
-  @Patch(':queueId/invite')
-  @Roles(Role.TA, Role.PROFESSOR)
-  async editQueueInvite(
-    @Param('queueId', ParseIntPipe) queueId: number,
-    @Body() body: QueueInviteParams,
-    @Res() res: Response,
-  ): Promise<Response<void>> {
-    try {
-      await this.queueService.editQueueInvite(queueId, body);
-      res.status(HttpStatus.OK).send();
-      return;
-    } catch (err) {
-      throw err;
-    }
-  }
 }
