+++ packages/server/test/queue.integration.ts
@@ -1237,226 +1237,4 @@ describe('Queue Integration', () => {
         .expect(400);
     });
   });
-
-  describe('POST /queues/:id/invite', () => {
-    it('returns 401 when the user is not logged in', async () => {
-      await supertest().post(`/queues/1/invite`).expect(401);
-    });
-    it('Creates a QueueInvite entity for the queue', async () => {
-      const course = await CourseFactory.create();
-      const ta = await TACourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-
-      await supertest({ userId: ta.userId })
-        .post(`/queues/${queue.id}/invite`)
-        .expect(201);
-
-      const invite = await QueueInviteModel.findOne({
-        queueId: queue.id,
-      });
-      expect(invite).toBeTruthy();
-      expect(invite.queueId).toEqual(queue.id);
-    });
-    it('Does not allow students to create invites', async () => {
-      const course = await CourseFactory.create();
-      const stu = await StudentCourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-
-      await supertest({ userId: stu.userId })
-        .post(`/queues/${queue.id}/invite`)
-        .expect(403);
-    });
-    it('Does not allow creating invites for non-existent queues', async () => {
-      const ta = await TACourseFactory.create({
-        user: await UserFactory.create(),
-      });
-
-      await supertest({ userId: ta.userId })
-        .post(`/queues/999/invite`)
-        .expect(404);
-    });
-    it('Only allows a queue to have a single invite', async () => {
-      const course = await CourseFactory.create();
-      const ta = await TACourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-
-      await supertest({ userId: ta.userId }).post(`/queues/${queue.id}/invite`);
-
-      await supertest({ userId: ta.userId })
-        .post(`/queues/${queue.id}/invite`)
-        .expect(400);
-    });
-  });
-  describe('DELETE /queues/:id/invite', () => {
-    it('returns 401 when the user is not logged in', async () => {
-      await supertest().delete(`/queues/1/invite`).expect(401);
-    });
-    it('Deletes a QueueInvite entity for the queue', async () => {
-      const course = await CourseFactory.create();
-      const ta = await TACourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-      await QueueInviteFactory.create({ queue: queue });
-
-      await supertest({ userId: ta.userId })
-        .delete(`/queues/${queue.id}/invite`)
-        .expect(204);
-
-      const invite = await QueueInviteModel.findOne({
-        queueId: queue.id,
-      });
-      expect(invite).toBeFalsy();
-    });
-    it('Does not allow students to delete invites', async () => {
-      const course = await CourseFactory.create();
-      const stu = await StudentCourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-      await QueueInviteFactory.create({ queueId: queue.id });
-
-      await supertest({ userId: stu.userId })
-        .delete(`/queues/${queue.id}/invite`)
-        .expect(403);
-    });
-    it('Does not allow deleting invites for non-existent queues', async () => {
-      const ta = await TACourseFactory.create({
-        user: await UserFactory.create(),
-      });
-
-      await supertest({ userId: ta.userId })
-        .delete(`/queues/999/invite`)
-        .expect(404);
-    });
-  });
-  describe('PATCH /queues/:id/invite', () => {
-    it('returns 401 when the user is not logged in', async () => {
-      await supertest().patch(`/queues/1/invite`).expect(401);
-    });
-    it('Updates a QueueInvite entity for the queue', async () => {
-      const course = await CourseFactory.create();
-      const ta = await TACourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-      const invite = await QueueInviteFactory.create({ queue: queue });
-
-      const newInvite: QueueInviteParams = {
-        queueId: queue.id,
-        QRCodeEnabled: true,
-        isQuestionsVisible: true,
-        willInviteToCourse: true,
-        inviteCode: 'newInviteCode',
-        QRCodeErrorLevel: 'M',
-      };
-
-      await supertest({ userId: ta.userId })
-        .patch(`/queues/${queue.id}/invite`)
-        .send(newInvite)
-        .expect(200);
-
-      const updatedInvite = await QueueInviteModel.findOne({
-        queueId: queue.id,
-      });
-
-      expect(updatedInvite).toBeTruthy();
-      expect(updatedInvite.queueId).toEqual(queue.id);
-      expect(updatedInvite.QRCodeEnabled).toEqual(newInvite.QRCodeEnabled);
-      expect(updatedInvite.isQuestionsVisible).toEqual(
-        newInvite.isQuestionsVisible,
-      );
-      expect(updatedInvite.willInviteToCourse).toEqual(
-        newInvite.willInviteToCourse,
-      );
-      expect(updatedInvite.inviteCode).toEqual(newInvite.inviteCode);
-      expect(updatedInvite.QRCodeErrorLevel).toEqual(
-        newInvite.QRCodeErrorLevel,
-      );
-    });
-    it('Does not allow students to update invites', async () => {
-      const course = await CourseFactory.create();
-      const stu = await StudentCourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-      const invite = await QueueInviteFactory.create({ queue: queue });
-
-      const newInvite: QueueInviteParams = {
-        queueId: queue.id,
-        QRCodeEnabled: true,
-        isQuestionsVisible: true,
-        willInviteToCourse: true,
-        inviteCode: 'newInviteCode',
-        QRCodeErrorLevel: 'M',
-      };
-
-      await supertest({ userId: stu.userId })
-        .patch(`/queues/${queue.id}/invite`)
-        .send(newInvite)
-        .expect(403);
-    });
-    it('Does not allow updating invites for non-existent queues', async () => {
-      const ta = await TACourseFactory.create({
-        user: await UserFactory.create(),
-      });
-
-      const newInvite: QueueInviteParams = {
-        queueId: 999,
-        QRCodeEnabled: true,
-        isQuestionsVisible: true,
-        willInviteToCourse: true,
-        inviteCode: 'newInviteCode',
-        QRCodeErrorLevel: 'M',
-      };
-
-      await supertest({ userId: ta.userId })
-        .patch(`/queues/999/invite`)
-        .send(newInvite)
-        .expect(404);
-    });
-    it('Does not allow updating invites with invalid data', async () => {
-      const course = await CourseFactory.create();
-      const ta = await TACourseFactory.create({
-        course: course,
-        user: await UserFactory.create(),
-      });
-      const queue = await QueueFactory.create({ course });
-      const invite = await QueueInviteFactory.create({ queue: queue });
-
-      const invalidInvite = "I'm not an object";
-
-      await supertest({ userId: ta.userId })
-        .patch(`/queues/${queue.id}/invite`)
-        .send(invalidInvite)
-        .expect(400);
-
-      const invalidInvite2 = {
-        queueId: queue.id,
-        QRCodeEnabled: 'A',
-        isQuestionsVisible: 'B',
-        willInviteToCourse: 'C',
-        inviteCode: 1,
-        QRCodeErrorLevel: 2,
-      };
-
-      await supertest({ userId: ta.userId })
-        .patch(`/queues/${queue.id}/invite`)
-        .send(invalidInvite2)
-        .expect(400);
-    });
-  });
 });
